+++ src/App.js
@@ -1,6 +1,6 @@
 import { BrowserRouter, Routes, Route } from 'react-router-dom';
 import './App.css';
-import Rockets from './components/Rockets';
+import Rockets from './components/Rockets/Rockets';
 
 function App() {
   return (
