+++ src/components/Rockets/Rockets.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import NavBar from './NavBar';
+import NavBar from '../Layout/NavBar';
 import RocketsList from './RocketsList';
 
 const Rockets = () => (
